General design
Look and Feel. You guys might need to build a path to the look and feel.
I uploaded respective jar files. If it doesn't work, let me know and I
can send screenshots of it first, and if you like it, I will continue to
work on it until we meet up on Tuesday to finalize the design.